Better precision for Context::param_info_to_json()
